TL change the doSum code for wild range

+++ Git_Project01/src/main/java/com/nirmal/service/Calculater.java
@@ -2,7 +2,7 @@ package com.nirmal.service;
 
 public class Calculater 
 {
-	public int doSum(int x,int y)
+	public double doSum(double x,double y)
 	{
 		return x+y;
 	}
